backend/python_server/requirements.txt
Flask==3.0.3
requests==2.32.3
flask-cors==4.0.1
python-dotenv==1.0.0
imgkit==1.0.0